calculator 클래스 파일 분리 및 적용
