+++ index.html
@@ -1,39 +1,43 @@
 <!DOCTYPE html>
 <html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Calculator</title>
-        <link rel="stylesheet" type="text/css" href="src/css/index.css" />
-    </head>
-    <body>
-        <div id="app">
-            <div class="calculator">
-                <h1 id="total">0</h1>
-                <div class="digits flex">
-                    <button class="digit">9</button>
-                    <button class="digit">8</button>
-                    <button class="digit">7</button>
-                    <button class="digit">6</button>
-                    <button class="digit">5</button>
-                    <button class="digit">4</button>
-                    <button class="digit">3</button>
-                    <button class="digit">2</button>
-                    <button class="digit">1</button>
-                    <button class="digit">0</button>
-                </div>
-                <div class="modifiers subgrid">
-                    <button class="modifier">AC</button>
-                </div>
-                <div class="operations subgrid">
-                    <button class="operation">/</button>
-                    <button class="operation">X</button>
-                    <button class="operation">-</button>
-                    <button class="operation">+</button>
-                    <button class="operation" id="equal">=</button>
-                </div>
+
+<head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Calculator</title>
+    <link rel="stylesheet" type="text/css" href="src/css/index.css" />
+</head>
+
+<body>
+    <div id="app">
+        <div class="calculator">
+            <h1 id="total">0</h1>
+            <div class="digits flex">
+                <button class="digit">9</button>
+                <button class="digit">8</button>
+                <button class="digit">7</button>
+                <button class="digit">6</button>
+                <button class="digit">5</button>
+                <button class="digit">4</button>
+                <button class="digit">3</button>
+                <button class="digit">2</button>
+                <button class="digit">1</button>
+                <button class="digit">0</button>
+            </div>
+            <div class="modifiers subgrid">
+                <button class="modifier">AC</button>
+            </div>
+            <div class="operations subgrid">
+                <button class="operation">/</button>
+                <button class="operation">X</button>
+                <button class="operation">-</button>
+                <button class="operation">+</button>
+                <button class="operation" id="equal">=</button>
             </div>
         </div>
-        <script src="/src/js/index.js"></script>
-    </body>
+    </div>
+    <script type="module" src="/src/js/index.js"></script>
+    <script type="module" src="/src/js/calculator.js"></script>
+</body>
+
 </html>
